blog-dev-005-refactor-controller-service - add title in post html, simplify post all controller, fix other html

+++ springBootBlog/src/main/java/com/yen/mdblog/entity/Po/Post.java
@@ -39,7 +39,7 @@ public class Post implements Serializable {
 	//@ManyToOne
 	//@JoinColumn(name = "author_id")
 	@Column
-	private Long authorId;
+	private Integer authorId;
 
 	@Column
 	@Convert(converter = DateTimeUtil.class)
